requirements.txt
streamlit
ffmpeg-python
pathlib